+++ src/test/java/de/uni_jena/cs/fusion/abecto/model/ModelsTest.java
@@ -15,22 +15,99 @@
  */
 package de.uni_jena.cs.fusion.abecto.model;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.ByteArrayInputStream;
 
+import org.apache.jena.rdf.model.Model;
+import org.apache.jena.rdf.model.ModelFactory;
+import org.apache.jena.rdf.model.Resource;
+import org.apache.jena.rdf.model.ResourceFactory;
+import org.apache.jena.riot.Lang;
+import org.apache.jena.riot.RDFDataMgr;
 import org.junit.jupiter.api.Test;
 
 public class ModelsTest {
 
-	@Test
-	public void testGetEmptyOntModel() {
-		assertFalse(Models.getEmptyOntModel().listStatements().hasNext());
+	private static Model modelOf(String ontology, Lang lang) {
+		Model model = ModelFactory.createDefaultModel();
+		RDFDataMgr.read(model, new ByteArrayInputStream(ontology.getBytes()), lang);
+		return model;
 	}
 
 	@Test
 	public void loadVeryShortOntologies() throws Exception {
 		Models.read(new ByteArrayInputStream(("@prefix : <http://example.org/>.\n:s :p :o.").getBytes()));
+	}
+
+	@Test
+	public void readOntologyIri() {
+		Resource expected = ResourceFactory.createResource("http://example.org/");
+		assertEquals(expected, Models.readOntologyIri(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2002/07/owl#Ontology> ."//
+				, Lang.TTL)).get());
+		assertEquals(expected, Models.readOntologyIri(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2004/02/skos/core#ConceptScheme> ."//
+				, Lang.TTL)).get());
+		assertTrue(Models.readOntologyIri(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2002/07/owl#Class> ."//
+				, Lang.TTL)).isEmpty());
+	}
+
+	@Test
+	public void readVersion() {
+		assertEquals("2.7.3", Models.readVersion(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2002/07/owl#Ontology> ;"//
+						+ "<http://www.w3.org/2002/07/owl#versionInfo> \"2.7.3\" ."//
+				, Lang.TTL)).get());
+	}
+
+	@Test
+	public void readVersionDateTime() {
+		assertEquals("2020-07-14", Models.readVersionDateTime(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2002/07/owl#Ontology> ;"//
+						+ "<http://purl.org/dc/terms/modified> \"2020-07-14\" ;"//
+						+ "<http://purl.org/dc/terms/available> \"2020-07-13\" ;"//
+						+ "<http://purl.org/dc/terms/created> \"2020-07-13\" ;"//
+						+ "<http://purl.org/dc/terms/date> \"2020-07-13\" ;"//
+						+ "<http://purl.org/dc/elements/1.1/date> \"2020-07-13\" ."//
+				, Lang.TTL)).get());
+		assertEquals("2020-07-14", Models.readVersionDateTime(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2002/07/owl#Ontology> ;"//
+						+ "<http://purl.org/dc/terms/available> \"2020-07-14\" ;"//
+						+ "<http://purl.org/dc/terms/created> \"2020-07-13\" ;"//
+						+ "<http://purl.org/dc/terms/date> \"2020-07-13\" ;"//
+						+ "<http://purl.org/dc/elements/1.1/date> \"2020-07-13\" ."//
+				, Lang.TTL)).get());
+		assertEquals("2020-07-14", Models.readVersionDateTime(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2002/07/owl#Ontology> ;"//
+						+ "<http://purl.org/dc/terms/created> \"2020-07-14\" ;"//
+						+ "<http://purl.org/dc/terms/date> \"2020-07-13\" ;"//
+						+ "<http://purl.org/dc/elements/1.1/date> \"2020-07-13\" ."//
+				, Lang.TTL)).get());
+		assertEquals("2020-07-14", Models.readVersionDateTime(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2002/07/owl#Ontology> ;"//
+						+ "<http://purl.org/dc/terms/date> \"2020-07-14\" ;"//
+						+ "<http://purl.org/dc/elements/1.1/date> \"2020-07-13\" ."//
+				, Lang.TTL)).get());
+		assertEquals("2020-07-14", Models.readVersionDateTime(modelOf(//
+				"<http://example.org/> a <http://www.w3.org/2002/07/owl#Ontology> ;"//
+						+ "<http://purl.org/dc/elements/1.1/date> \"2020-07-14\" ."//
+				, Lang.TTL)).get());
+	}
+
+	@Test
+	public void readVersionIri() {
+		assertEquals(ResourceFactory.createResource("http://example.org/2.7.3/"), Models.readVersionIri(modelOf(//
+				"<http://example.org/> <http://www.w3.org/2002/07/owl#versionIRI> <http://example.org/2.7.3/> ."//
+				, Lang.TTL)).get());
+	}
+
+	@Test
+	public void testGetEmptyOntModel() {
+		assertFalse(Models.getEmptyOntModel().listStatements().hasNext());
 	}
 
 }